fix(slack): bind HMACs per installation, dedup modal after validation, update all fanned-out cards; satisfy CI gates (PR #1910)
Addresses the Codex review findings and both CI gate failures on PR #1910.

- P1 — installation-bound signature verification: the events and
  interactions endpoints now verify the Slack HMAC against the signing
  secret(s) of the installation the payload CLAIMS (team_id + api_app_id),
  not the union of all configured secrets — a payload for connection B can
  no longer be admitted on connection A's secret. Payloads matching no
  configured installation (or carrying none, e.g. url_verification) fall
  back to the configured-secret set and are still rejected by installation
  validation before anything runs. New cross-secret test proves a payload
  claiming app B signed with app A's secret is refused while the genuine
  signature is accepted end-to-end.
- P2 — modal dedup after validation: the rework view-id dedup key is now
  recorded only when a submission is fully valid and about to dispatch.
  Recording earlier burned the id on an empty-reason validation error
  (Slack keeps the same modal open), silently dropping the corrected
  resubmit. The round-trip test now reuses one view id across the rejected
  and corrected submissions.
- P2 — fan-out card updates: ApprovalMessageMap gains message_deliveries
  (one entry per channel+recipient, re-sends replace their own entry;
  legacy single-record files fall back transparently) and gate-decision
  updates now edit EVERY fanned-out card instead of only the last-recorded
  one, so no stale actionable card survives a decision.
- CI file-size gate: split the TAN-763..766 ingress tests into
  http/tests/slack_events_governance.rs and the channel approval-card
  update block into routines_automations_parts/part08.rs, bringing both
  touched files back inside the 2000-line hard cap.
- DCO: this and all prior branch commits are signed off.
