+++ crates/tandem-server/src/http/tests/mod.rs
@@ -48,6 +48,7 @@ pub(super) mod routines;
 pub(super) mod sessions;
 pub(super) mod setup_understanding;
 pub(super) mod slack_events;
+pub(super) mod slack_events_governance;
 pub(super) mod stateful_runtime_hardening;
 pub(super) mod stateful_runtime_observability_contracts;
 pub(super) mod task_intake;
